+++ frontend/packages/erd-core/src/features/erd/components/ERDRenderer/CommandPalette/CommandPaletteOptions/TableDetailOptions.test.tsx
@@ -26,7 +26,7 @@ afterEach(() => {
 })
 
 const mockSetCommandPaletteDialogOpen = vi.fn()
-const mockSelectTable = vi.fn()
+const mockRevealTable = vi.fn()
 const mockWindowOpen = vi.fn()
 
 const originalUseCommandPaletteOrThrow =
@@ -45,7 +45,7 @@ vi.spyOn(UseTableSelection, 'useTableSelection').mockImplementation(() => {
   const original = originalUseTableSelection()
   return {
     ...original,
-    selectTable: mockSelectTable,
+    revealTable: mockRevealTable,
   }
 })
 vi.spyOn(window, 'open').mockImplementation(mockWindowOpen)
@@ -151,7 +151,7 @@ describe('mouse interactions', () => {
 
       await user.click(screen.getByRole('link', { name: elementName }))
 
-      expect(mockSelectTable).toHaveBeenCalled()
+      expect(mockRevealTable).toHaveBeenCalled()
       expect(mockSetCommandPaletteDialogOpen).toHaveBeenCalledWith(false)
       expect(window.location.hash).toBe(hash)
     })
@@ -166,7 +166,7 @@ describe('mouse interactions', () => {
       await user.click(screen.getByRole('link', { name: elementName }))
       await user.keyboard('{/Meta}')
 
-      expect(mockSelectTable).not.toHaveBeenCalled()
+      expect(mockRevealTable).not.toHaveBeenCalled()
       expect(mockSetCommandPaletteDialogOpen).not.toHaveBeenCalled()
 
       // FIXME: jsdom doesn't implement behavior of ⌘ + click to open a link in a new tab, but it changes the URL of the current window
@@ -214,7 +214,7 @@ describe('keyboard interactions', () => {
 
       await user.keyboard('{Enter}')
 
-      expect(mockSelectTable).toHaveBeenCalledWith({
+      expect(mockRevealTable).toHaveBeenCalledWith({
         displayArea: 'main',
         tableId: 'users',
       })
@@ -236,7 +236,7 @@ describe('keyboard interactions', () => {
       expect(mockWindowOpen).toHaveBeenCalledWith(`?active=users${hash}`)
 
       // other functions are not called
-      expect(mockSelectTable).not.toHaveBeenCalled()
+      expect(mockRevealTable).not.toHaveBeenCalled()
       expect(mockSetCommandPaletteDialogOpen).not.toHaveBeenCalled()
     })
   })
@@ -254,7 +254,7 @@ describe('keyboard interactions', () => {
     await user.keyboard('{Meta>}{Enter}{/Meta}')
 
     expect(mockWindowOpen).not.toHaveBeenCalled()
-    expect(mockSelectTable).not.toHaveBeenCalled()
+    expect(mockRevealTable).not.toHaveBeenCalled()
     expect(mockSetCommandPaletteDialogOpen).not.toHaveBeenCalled()
   })
 })