+++ frontend/packages/erd-core/src/features/erd/components/ERDRenderer/CommandPalette/CommandPaletteOptions/TableOptions.test.tsx
@@ -21,7 +21,7 @@ afterEach(() => {
 })
 
 const mockSetCommandPaletteDialogOpen = vi.fn()
-const mockSelectTable = vi.fn()
+const mockRevealTable = vi.fn()
 const mockWindowOpen = vi.fn()
 
 const originalUseCommandPaletteOrThrow =
@@ -40,7 +40,7 @@ vi.spyOn(UseTableSelection, 'useTableSelection').mockImplementation(() => {
   const original = originalUseTableSelection()
   return {
     ...original,
-    selectTable: mockSelectTable,
+    revealTable: mockRevealTable,
   }
 })
 vi.spyOn(window, 'open').mockImplementation(mockWindowOpen)
@@ -91,7 +91,7 @@ describe('mouse interactions', () => {
 
     await user.click(screen.getByRole('link', { name: 'follows' }))
 
-    expect(mockSelectTable).toHaveBeenCalled()
+    expect(mockRevealTable).toHaveBeenCalled()
     expect(mockSetCommandPaletteDialogOpen).toHaveBeenCalledWith(false)
   })
 
@@ -103,7 +103,7 @@ describe('mouse interactions', () => {
     await user.click(screen.getByRole('link', { name: 'follows' }))
     await user.keyboard('{/Meta}')
 
-    expect(mockSelectTable).not.toHaveBeenCalled()
+    expect(mockRevealTable).not.toHaveBeenCalled()
     expect(mockSetCommandPaletteDialogOpen).not.toHaveBeenCalled()
   })
 })
@@ -117,7 +117,7 @@ describe('keyboard interactions', () => {
 
     await user.keyboard('{Enter}')
 
-    expect(mockSelectTable).toHaveBeenCalledWith({
+    expect(mockRevealTable).toHaveBeenCalledWith({
       displayArea: 'main',
       tableId: 'users',
     })
@@ -138,7 +138,7 @@ describe('keyboard interactions', () => {
     expect(mockWindowOpen).toHaveBeenCalledWith('?active=users')
 
     // other functions are not called
-    expect(mockSelectTable).not.toHaveBeenCalled()
+    expect(mockRevealTable).not.toHaveBeenCalled()
     expect(mockSetCommandPaletteDialogOpen).not.toHaveBeenCalled()
   })
 
@@ -152,7 +152,7 @@ describe('keyboard interactions', () => {
     await user.keyboard('{Meta>}{Enter}{/Meta}')
 
     expect(mockWindowOpen).not.toHaveBeenCalled()
-    expect(mockSelectTable).not.toHaveBeenCalled()
+    expect(mockRevealTable).not.toHaveBeenCalled()
     expect(mockSetCommandPaletteDialogOpen).not.toHaveBeenCalled()
   })
 })